Move to ref and hooks based implementation
