+++ packages/react-events/src/FocusScope.js
@@ -9,8 +9,17 @@
 import type {
   ReactResponderEvent,
   ReactResponderContext,
+  ReactEventComponentInstance,
 } from 'shared/ReactTypes';
 import React from 'react';
+import {
+  isFiberSuspenseAndTimedOut,
+  getSuspenseFallbackChild,
+} from 'react-reconciler/src/ReactFiberEvents';
+import {HostComponent} from 'shared/ReactWorkTags';
+import type {Fiber} from 'react-reconciler/src/ReactFiber';
+import ReactSharedInternals from 'shared/ReactSharedInternals';
+import invariant from 'shared/invariant';
 
 type FocusScopeProps = {
   autoFocus: boolean,
@@ -23,8 +32,109 @@ type FocusScopeState = {
   currentFocusedNode: null | HTMLElement,
 };
 
+type FocusManagerOptions = {
+  from?: HTMLElement,
+  tabbable?: boolean,
+  wrap?: boolean,
+};
+
+type FocusManager = {
+  focusNext(opts: FocusManagerOptions): HTMLElement,
+  focusPrevious(opts: FocusManagerOptions): HTMLElement,
+};
+
 const targetEventTypes = [{name: 'keydown', passive: false}];
 const rootEventTypes = [{name: 'focus', passive: true}];
+
+function getFocusableElementsInScope(
+  eventComponentInstance: ReactEventComponentInstance,
+  tabbable?: boolean,
+  searchNode?: Element,
+): Array<HTMLElement> {
+  const focusableElements = [];
+  const child = ((eventComponentInstance.currentFiber: any): Fiber).child;
+
+  if (child !== null) {
+    collectFocusableElements(child, focusableElements, tabbable, searchNode);
+  }
+  return focusableElements;
+}
+
+function collectFocusableElements(
+  node: Fiber,
+  focusableElements: Array<HTMLElement>,
+  tabbable?: boolean,
+  searchNode?: Element,
+): void {
+  if (isFiberSuspenseAndTimedOut(node)) {
+    const fallbackChild = getSuspenseFallbackChild(node);
+    if (fallbackChild !== null) {
+      collectFocusableElements(
+        fallbackChild,
+        focusableElements,
+        tabbable,
+        searchNode,
+      );
+    }
+  } else {
+    if (
+      isFiberHostComponentFocusable(node, tabbable) ||
+      node.stateNode === searchNode
+    ) {
+      focusableElements.push(node.stateNode);
+    } else {
+      const child = node.child;
+
+      if (child !== null) {
+        collectFocusableElements(
+          child,
+          focusableElements,
+          tabbable,
+          searchNode,
+        );
+      }
+    }
+  }
+  const sibling = node.sibling;
+
+  if (sibling !== null) {
+    collectFocusableElements(sibling, focusableElements, tabbable, searchNode);
+  }
+}
+
+function isFiberHostComponentFocusable(
+  fiber: Fiber,
+  tabbable?: boolean,
+): boolean {
+  if (fiber.tag !== HostComponent) {
+    return false;
+  }
+  const {type, memoizedProps} = fiber;
+  if ((tabbable && memoizedProps.tabIndex === -1) || memoizedProps.disabled) {
+    return false;
+  }
+  const minTabIndex = tabbable ? 0 : -1;
+  if (
+    (memoizedProps.tabIndex != null && memoizedProps.tabIndex >= minTabIndex) ||
+    memoizedProps.contentEditable === true
+  ) {
+    return true;
+  }
+  if (type === 'a' || type === 'area') {
+    return !!memoizedProps.href && memoizedProps.rel !== 'ignore';
+  }
+  if (type === 'input') {
+    return memoizedProps.type !== 'hidden' && memoizedProps.type !== 'file';
+  }
+  return (
+    type === 'button' ||
+    type === 'textarea' ||
+    type === 'object' ||
+    type === 'select' ||
+    type === 'iframe' ||
+    type === 'embed'
+  );
+}
 
 function focusElement(element: ?HTMLElement) {
   if (element != null) {
@@ -34,11 +144,76 @@ function focusElement(element: ?HTMLElement) {
   }
 }
 
+function moveFocusInScope(
+  scope: ReactEventComponentInstance,
+  node: Element,
+  backwards: boolean,
+  options: FocusManagerOptions = {},
+) {
+  let elements = getFocusableElementsInScope(scope, !!options.tabbable, node);
+  if (elements.length === 0) {
+    return null;
+  }
+
+  const position = elements.indexOf(node);
+  const lastPosition = elements.length - 1;
+  let nextElement = null;
+
+  if (backwards) {
+    if (position === 0) {
+      if (options.wrap) {
+        nextElement = elements[lastPosition];
+      } else {
+        // Out of bounds
+        return null;
+      }
+    } else {
+      nextElement = elements[position - 1];
+    }
+  } else {
+    if (position === lastPosition) {
+      if (options.wrap) {
+        nextElement = elements[0];
+      } else {
+        // Out of bounds
+        return null;
+      }
+    } else {
+      nextElement = elements[position + 1];
+    }
+  }
+
+  if (nextElement) {
+    focusElement(nextElement);
+    return nextElement;
+  }
+
+  return null;
+}
+
+function moveFocus(
+  scope: ReactEventComponentInstance,
+  options: FocusManagerOptions = {},
+  backwards: boolean,
+) {
+  let node = options.from;
+  if (!node) {
+    let doc = typeof document !== 'undefined' ? document : null;
+    node = doc && doc.activeElement;
+  }
+
+  if (!node) {
+    return;
+  }
+
+  return moveFocusInScope(scope, node, backwards, options);
+}
+
 function getFirstFocusableElement(
   context: ReactResponderContext,
   state: FocusScopeState,
 ): ?HTMLElement {
-  const elements = context.getFocusableElementsInScope();
+  const elements = getFocusableElementsInScope(context.getCurrentInstance());
   if (elements.length > 0) {
     return elements[0];
   }
@@ -75,22 +250,19 @@ const FocusScopeResponder = {
           return;
         }
 
-        const nextElement = context.moveFocusInScope(focusedElement, shiftKey, {
-          tabbable: true,
-          wrap: props.contain,
-        });
+        const nextElement = moveFocusInScope(
+          context.getCurrentInstance(),
+          focusedElement,
+          shiftKey,
+          {
+            tabbable: true,
+            wrap: props.contain,
+          },
+        );
 
         if (!nextElement) {
           context.continueLocalPropagation();
-          return;
-        }
-
-        // If this element is possibly inside the scope of another
-        // FocusScope responder or is out of bounds, then we release ownership.
-        if (nextElement !== null) {
-          if (!context.isTargetWithinEventResponderScope(nextElement)) {
-            context.continueLocalPropagation();
-          }
+        } else {
           state.currentFocusedNode = nextElement;
           ((nativeEvent: any): KeyboardEvent).preventDefault();
         }
@@ -142,7 +314,40 @@ const FocusScopeResponder = {
   },
 };
 
-export default React.unstable_createEventComponent(
+export const FocusScopeEventComponent = React.unstable_createEventComponent(
   FocusScopeResponder,
   'FocusScope',
 );
+
+const FocusScopeContext: React.Context<FocusManager> = React.createContext();
+
+export const FocusScope = React.forwardRef(
+  (props: FocusScopeProps, ref: React.Ref<FocusManager>) => {
+    let fiber = ReactSharedInternals.ReactCurrentOwner.current;
+    let focusManager = {
+      focusNext(options: FocusManagerOptions = {}) {
+        // forwardRef -> context provider -> event component -> responder
+        return moveFocus(fiber.child.child.stateNode, options, false);
+      },
+      focusPrevious(options: FocusManagerOptions = {}) {
+        return moveFocus(fiber.child.child.stateNode, options, true);
+      },
+    };
+
+    React.useImperativeHandle(ref, () => focusManager);
+
+    return React.createElement(FocusScopeContext.Provider, {
+      value: focusManager,
+      children: React.createElement(FocusScopeEventComponent, props),
+    });
+  },
+);
+
+export function useFocusManager(): FocusManager {
+  let focusManager = React.useContext(FocusScopeContext);
+  invariant(
+    focusManager != null,
+    'Tried to call useFocusManager outside of a FocusScope subtree.',
+  );
+  return focusManager;
+}
